+++ examples/data-sources/criblio_source/data-source.tf
@@ -1,0 +1,3 @@
+data "criblio_source" "my_source" {
+  group_id = "...my_group_id..."
+}