+++ examples/resources/criblio_deploy/resource.tf
@@ -1,0 +1,4 @@
+resource "criblio_deploy" "my_deploy" {
+  id      = "...my_id..."
+  version = "...my_version..."
+}