+++ examples/resources/criblio_destination/import.sh
@@ -1,0 +1,1 @@
+terraform import criblio_destination.my_criblio_destination ""