+++ examples/resources/criblio_pack/resource.tf
@@ -1,5 +1,10 @@
 resource "criblio_pack" "my_pack" {
-  filename = "...my_filename..."
-  group_id = "...my_group_id..."
-  id       = "...my_id..."
+  description  = "...my_description..."
+  disabled     = true
+  display_name = "...my_display_name..."
+  filename     = "...my_filename..."
+  group_id     = "...my_group_id..."
+  id           = "...my_id..."
+  source       = "...my_source..."
+  version      = "...my_version..."
 }