+++ examples/resources/criblio_source/import.sh
@@ -1,0 +1,1 @@
+terraform import criblio_source.my_criblio_source ""