+++ examples/stream-syslog-to-lake/main.tf
@@ -135,6 +135,28 @@ resource "criblio_pack" "syslog_pack" {
   id       = "syslog-processing"
   group_id = criblio_group.syslog_worker_group.id
   filename = "cribl-palo-alto-networks-source-1.0.0.crbl" 
+  description  = "Pack for syslog processing"
+  disabled     = false
+  display_name = "Pack for syslog processing"
+  version      = "1.0.0"
+}
+
+# Commit and Deploy Configuration
+data "criblio_config_version" "my_configversion" {
+  id = "syslog-workers"
+  depends_on = [criblio_commit.my_commit]
+}
+
+resource "criblio_commit" "my_commit" {
+  effective = true
+  group   = "syslog-workers"
+  message = "test"
+  depends_on = [criblio_source.syslog_source, criblio_destination.cribl_lake, criblio_pack.syslog_pack]
+}
+
+resource "criblio_deploy" "my_deploy" {
+  id      = "syslog-workers"
+  version = data.criblio_config_version.my_configversion.items[0]
 }
 
 # Outputs