+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2024-03-21
+
+### Added
+- Support for managing Cribl commits
+- Support for managing Cribl deployments
+- Data source support for querying Cribl config versions
+
 ## [1.0.0] - 2024-03-21
 
 ### Added
@@ -19,4 +26,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Data source support for querying Cribl pack pipelines
 - Data source support for querying Cribl pipelines
 
+[1.0.8]: https://github.com/criblio/terraform-provider-criblio/releases/tag/v1.0.8
 [1.0.0]: https://github.com/criblio/terraform-provider-criblio/releases/tag/v1.0.0